requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
httpx==0.25.2
aiohttp==3.9.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
torch==2.4.1
networkx==3.2.1
numpy==1.26.4
